Ready to Deploy

+++ src/Componets/DashboardComponents/ParkingElements.js
@@ -27,28 +27,30 @@ const ParkingElement = () => {
 
   // Function to delete a parking lot
   const deleteLot = async (id) => {
-    console.log("Attempting to delete the parking lot with ID:", id);
-    try {
-      const response = await fetch(
-        `https://localhost:7040/api/ParkingSpace/DeleteParkingSpace/${id}`,
-        {
-          method: "DELETE",
-          headers: {
-            "Content-Type": "application/json",
-            Authorization: `Bearer ${localStorage.getItem("token")}`,
-          },
-        }
-      );
+    if (window.confirm("Are you sure you want to delete this parking lot?")) {
+      console.log("Attempting to delete the parking lot with ID:", id);
+      try {
+        const response = await fetch(
+          `https://localhost:7040/api/ParkingSpace/DeleteParkingSpace/${id}`,
+          {
+            method: "DELETE",
+            headers: {
+              "Content-Type": "application/json",
+              Authorization: `Bearer ${localStorage.getItem("token")}`,
+            },
+          }
+        );
 
-      if (response.ok) {
-        console.log("Parking lot deleted successfully");
-        fetchParkingLots(); // Refetch the data after deletion
-      } else {
-        const errorText = await response.json();
-        console.error("Failed to delete parking lot:", errorText);
+        if (response.ok) {
+          console.log("Parking lot deleted successfully");
+          setParkingLot(parkingLot.filter((lot) => lot.spaceId !== id)); // Update the state
+        } else {
+          const errorText = await response.json();
+          console.error("Failed to delete parking lot:", errorText);
+        }
+      } catch (error) {
+        console.error("Error deleting parking lot:", error);
       }
-    } catch (error) {
-      console.error("Error deleting parking lot:", error);
     }
   };
 
@@ -66,7 +68,7 @@ const ParkingElement = () => {
                   <th>Available Space</th>
                   <th>Location</th>
                   <th className="th_5">Price</th>
-                  <th>Actions</th>
+                  <th className="th_6">Actions</th>
                 </tr>
               </thead>
               <tbody>
